Scripted execution updates
- Support examples with multiple cases.
- Improved help functionality.
- Support a common data folder.
- Reduced data copied around by introducing symbolic links to meshes.

+++ run.sh
@@ -1,57 +1,58 @@
 #!/bin/bash
-
 set -e
+# ============================================================================ #
+# Define the help function
+function help() {
+    echo -e "Execution of Neko cases."
+
+    echo -e "\n\e[4mUsage:\e[0m"
+    echo -e "  run.sh [options] [example]"
+
+    echo -e "\n\e[4mDescription:\e[0m"
+    echo -e "  This script works as a function which run the examples specified"
+    echo -e "  through the command line."
+    echo -e ""
+    echo -e "  The <EXAMPLE> input refers to the name or pattern of the case"
+    echo -e "  file. The examples folder is searched for the specified pattern."
+    echo -e "  If multiple matching case files are found, then all of them are"
+    echo -e "  run."
+    echo -e ""
+    echo -e "  See Readme for additional details."
+    echo -e ""
+
+    printf "\e[4mOptions:\e[0m\n"
+    printf "  -%-1s, --%-10s %-60s\n" "a" "all" "Run all journals available."
+    printf "  -%-1s, --%-10s %-60s\n" "c" "clean" "Clean artifacts from previous runs."
+    printf "  -%-1s, --%-10s %-60s\n" "d" "delete" "Delete previous runs."
+    printf "  -%-1s, --%-10s %-60s\n" "h" "help" "Print help."
+
+    printf "\n\e[4mAvailable case files:\e[0m\n"
+    for case in $(find $EPATH -name "*.case" 2>/dev/null); do
+        printf '  - %-s\n' ${case#$EPATH/}
+    done
+    exit 0
+}
+if [ $# -lt 1 ]; then help; fi
 
 # ============================================================================ #
 # User defined inputs.
+MAIN_DIR=$(dirname $(realpath $0))
 
 # Define all needed folders relative to the project folder. (without trailing /)
-EPATH="examples" # Examples scripts
-RPATH="results"  # Result export location
-LPATH="logs"     # Logging locations
-HPATH="scripts"  # Submission settings
+export EPATH="$MAIN_DIR/examples" # Examples scripts
+export RPATH="$MAIN_DIR/results"  # Result export location
+export LPATH="$MAIN_DIR/logs"     # Logging locations
+export HPATH="$MAIN_DIR/scripts"  # Submission settings
+export DPATH="$MAIN_DIR/data"     # Meshes
 
 # End of user inputs
-# ============================================================================ #
-# Help commands
-function help() {
-    echo "run.sh EXAMPLE [EXAMPLE ...]"
-    echo "  This script works as a function which run the examples specified"
-    echo "  through the command line."
-    echo ""
-    echo "  The <EXAMPLE> input refers to the name of the subfolder located in"
-    echo "  $EPATH. This folder should contain everything needed to run the "
-    echo "  example."
-    echo ""
-    echo "  See Readme for details."
-    echo ""
-    echo "Available examples;"
-    for example in $(find $EPATH -d 2>/dev/null); do
-        if [ -f $example/*.case ]; then
-            t_name=${example#$EPATH/}
-            printf '  - %-s\n' $t_name
-        fi
-    done
-}
-
-# Variable setups
-MAINDIR=$PWD
-
-export EPATH=$MAINDIR/$EPATH
-export RPATH=$MAINDIR/$RPATH
-export LPATH=$MAINDIR/$LPATH
-export HPATH=$MAINDIR/$HPATH
-
 # ============================================================================ #
 # Keywords
 
 # Empty input
-if [ -z "$1" ]; then
-    help && exit
-fi
+if [ $# -lt 1 ]; then help; fi
 
-examples=""
-
+case_files=""
 for in in $@; do
 
     # Opions and flags
@@ -59,25 +60,21 @@ for in in $@; do
         case "${in:2}" in
         "all") ALL=true ;;       # Run all examples available
         "clean") CLEAN=true ;;   # Clean logs
-        "help") HELP=true ;;     # Print help
+        "help") help ;;          # Print help
         "delete") DELETE=true ;; # Delete previous runs
         *) printf '  %-10s %-67s\n' "Invalid option:" "$in" && exit 1 ;;
         esac
-
-        if [ $HELP ]; then help && exit; fi
 
     elif [[ ${in:0:1} == "-" ]]; then
         for ((i = 1; i < ${#in}; i++)); do
             case "${in:$i:1}" in
             "a") ALL=true ;;    # Run all examples available
             "c") CLEAN=true ;;  # Clean logs
-            "h") HELP=true ;;   # Print help
+            "h") help ;;        # Print help
             "d") DELETE=true ;; # Delete previous runs
             *) printf '  %-10s %-67s\n' "Invalid option:" "${in:$i:1}" && exit 1 ;;
             esac
         done
-
-        if [ $HELP ]; then help && exit; fi
 
     # Ignore invalid inputs
     elif [[ -z $(ls $EPATH/$in 2>/dev/null) ]]; then
@@ -85,30 +82,20 @@ for in in $@; do
 
     # Extract the examples from the input
     elif [[ ! $ALL ]]; then
-        for example in $(find $EPATH/$in -type d); do
-
-            case $(find $example -maxdepth 1 -name "*.case" | wc -l) in
-            0) ;;
-            1) examples+="${example#$EPATH/} " ;;
-            *) printf "Multiple case files in: ${example#$EPATH/}\n" >&2 && exit 1 ;;
-            esac
-
+        for case in $(find $EPATH/$in -name "*.case"); do
+            case_files+="${case#$EPATH/} "
         done
     fi
 done
 
 if [ $ALL ]; then
-    examples=""
-    for example in $(find $EPATH/ -type d); do
-        case $(find $example -maxdepth 1 -name "*.case" | wc -l) in
-        0) ;;
-        1) examples+="${example#$EPATH/} " ;;
-        *) printf "Multiple case files in: ${example#$EPATH/}\n" >&2 && exit 1 ;;
-        esac
+    case_files=""
+    for case in $(find $EPATH/ -name "*.case"); do
+        case_files+="${case#$EPATH/} "
     done
 fi
 
-if [ ! "$examples" ]; then
+if [ ! "$case_files" ]; then
     exit
 fi
 
@@ -131,71 +118,73 @@ fi
 
 # Function for running the examples
 function Run() {
-    example=$1
-    cd $LPATH/$example
+    case=$1
+    echo "Launching case on local machine" 1>output.out
 
-    echo "Launching example on local machine" 1>output.out
-
-    # Run the example and pipe stdout and stderr to the log files
-    ./job_script.sh $example >output.out 2>error.err &
-    printf '  %-10s %-s\n' "Started:" "$example"
+    # Run the case and pipe stdout and stderr to the log files
+    ./job_script.sh $case >output.out 2>error.err &
+    printf '  %-10s %-s\n' "Started:" "$case"
     wait
-
-    cd $MAINDIR
 }
 
 # Function for submitting the examples
 function Submit() {
-    example=$1
-    cd $LPATH/$example
-
-    touch output.out
-    echo "Launching example on HPC" 1>output.out
+    case=$1
+    echo "Launching case on HPC" 1>output.out
 
     export BSUB_QUIET=Y
-    bsub -J $example -env "RPATH=$RPATH" <./job_script.sh
-    printf '  %-10s %-s\n' "Submitted:" "$example"
-
-    cd $MAINDIR
+    bsub -J $case -env "RPATH=$RPATH" <job_script.sh
+    printf '  %-10s %-s\n' "Submitted:" "$case"
 }
 
 # ============================================================================ #
 # Run the examples
 full_start=$(date +%s.%N)
 
-mkdir -p $RPATH $LPATH
-printf "\n\e[4mQueueing examples.\e[0m\n"
-for example in $examples; do
-    if [ ! -f $EPATH/$example/neko ]; then
-        printf '  \e[1;31m%-13s\e[m %-s\n' "Error:" "Not Compiled: $example" >&2
+printf "\n\e[4mQueueing case files.\e[0m\n"
+for case in $case_files; do
+    case_name=$(basename ${case%.case})
+    case_dir=$(dirname $case)
+
+    if [ ! -f $EPATH/$case_dir/neko ]; then
+        printf '  \e[1;31m%-13s\e[m %-s\n' "Error:" "Not Compiled: $case" >&2
         exit 1
     fi
 
-    # Create the folder needed
-    log=$LPATH/$example
-
-    # Cleanup old log files
-    if [ $CLEAN ]; then
-        rm -fr $log
+    # Define the name of the current exampel, if there are multiple cases in the
+    # same folder, we add the case name to the example name.
+    example=${case_dir#$EPATH/}
+    if [[ $(find $EPATH/$case_dir -name "*.case" | wc -l) > 1 ]]; then
+        example=$example/$case_name
     fi
-    mkdir -p $log
-    cd $log
+
+    # Setup the log folder
+    log=$LPATH/$example && mkdir -p $log
+    if [ $CLEAN ]; then rm -fr $log/*; fi
 
     # Remove old output and error files
-    rm -f output.out error.err
-    touch output.out error.err
+    rm -f $log/output.out $log/error.err
 
     # Determine if we have a HPC file
-    setting=$HPATH/${example}.sh
-    if [ ! -f $setting ]; then setting=$HPATH/${example%/*}.sh; fi
+    setting=$HPATH/${case}.sh
+    if [ ! -f $setting ]; then setting=$HPATH/${case%/*}.sh; fi
     if [ ! -f $setting ]; then setting=$HPATH/default.sh; fi
 
-    cp -rt . $EPATH/$example/neko $EPATH/$example/*.case $EPATH/$example/*.nmsh
-    cp $setting ./job_script.sh
+    cp -ft $log $EPATH/$case $EPATH/$case_dir/neko
+    cp -f $setting $log/job_script.sh
+
+    # Create symbolic links to the mesh files to avoid copying massive files
+    for file in $(find $EPATH/$case_dir -name "*.nmsh" 2>/dev/null); do
+        ln -fs $file $log
+    done
+
+    if [ -d "$DPATH" ]; then ln -fs $DPATH $log; fi
 
     # Done with the setup
     # ======================================================================== #
     # Move to the directory submit the code and return
+
+    cd $log
 
     if [ "$(which bsub)" ]; then
         Submit $example
@@ -203,6 +192,7 @@ for example in $examples; do
         Run $example
     fi
 
+    cd $MAIN_DIR
 done
 
 if [ ! "$(which bsub)" ]; then
